set prob_t to just double

+++ src/include/prob_t.h
@@ -4,6 +4,6 @@
 #ifndef __ECOZ_PROB_T_H
 #define __ECOZ_PROB_T_H
 
-typedef long double prob_t;
+typedef double prob_t;
 
 #endif
